Skrimage 1 12 24 (#26)
* Doing cool stuff

Super auto now to grab white pixels

* Update SuperAutoRedBoardArmController.java

* Update SuperAutoRedBoardArmController.java

added all the actions for placing a pixel if it gets intaked at the spot. Still need to make the intake part better, it's just a simple move up and turn on the intake. Got some ideas but we need to test and I need sleep.

* Update Encoder Config RoadRunner

* Road Runner Tuning

* Update SampleMecanumDrive.java

* Update SuperAutoRedBoardArmController.java

* beep boop

auto stuffs, adjusting things due to testing

* Update auto

* update red working!

* Update Team Prop Detection

* Auto Fixes & Clean Up

* Merge Teleop Changes to skrimage Branch (#25)

* temp fix auto things

* Update With New drone launcher settings

* Update Fixes

* Skrimage Updates

* Auto Fixes

---------

Co-authored-by: JEEF-THE-LEEF <jacob.dulin@gmail.com>